+++ src/lib/server/dl_module.c
@@ -402,7 +402,7 @@ dl_module_t const *dl_module(dl_module_t const *parent, char const *name, dl_mod
 
 	if (!module_name) return NULL;
 
-	for (p = module_name, q = p + talloc_array_length(p) - 1; p < q; p++) *p = tolower(*p);
+	for (p = module_name, q = p + talloc_array_length(p) - 1; p < q; p++) *p = tolower((uint8_t) *p);
 
 	/*
 	 *	If the module's already been loaded, increment the reference count.
